Refactor WebAudio initialization
This moves the audio context creation into the user interaction
event handler for the mute/unmute button. Doing it this way fixes
a deprecated feature warning from Chrome. The warning was related
to API changes to prevent people from abusing autoplaying video
and audio.

This change also fixes a quirk where samples that were in progress
when the mute button was pressed would surprisingly resume playing
if you unmute the audio later. For playing recorded music, that
behavior would make sense. But, for a drum synth, it's better to
unmute with a fresh start with no sounds playing.