Cache ArrayBuffers for .flac sample files
This caches the ArrayBuffer objects for each drum sample .flac
file so they can be used repeatedly. This avoids the latency of
having to go through fetch() and the download cache.

The trick to getting this working is that the ArrayBuffer object
must be copied before it gets turned over to the audio playback
system. Copying the buffer takes some time. But, that's still a
big improvement over downloading the whole file again. So, this is
meaningful progress.